feat: improve repository error checking, add more overloads

+++ app/src/main/java/com/example/eventapp/repositories/EventRepository.java
@@ -47,10 +47,12 @@ public class EventRepository {
     }
 
     public Task<Void> updateEvent(Event event) {
+        String documentId = event.getDocumentId();
+        if (documentId == null) throw new NullPointerException("documentId is null - never set documentId");
+
         return eventCollection.document(event.getDocumentId()).set(event)
             .addOnCompleteListener(task -> {
                 if (task.isSuccessful()) {
-                    String documentId = event.getDocumentId();
                     Log.d(TAG, "updateEvent: success - ID: " + documentId);
                 } else {
                     Log.e(TAG, "updateEvent: fail", task.getException());
@@ -59,10 +61,12 @@ public class EventRepository {
     }
 
     public Task<Void> removeEvent(Event event) {
+        String documentId = event.getDocumentId();
+        if (documentId == null) throw new NullPointerException("documentId is null - never set documentId");
+
         return eventCollection.document(event.getDocumentId()).delete()
             .addOnCompleteListener(task -> {
                 if (task.isSuccessful()) {
-                    String documentId = event.getDocumentId();
                     Log.d(TAG, "removeEvent: success - ID: " + documentId);
                 } else {
                     Log.e(TAG, "removeEvent: fail", task.getException());
@@ -71,12 +75,31 @@ public class EventRepository {
     }
 
     public LiveData<List<Event>> getEventsOfOrganizerLiveData(String organizerId) {
-        MutableLiveData<List<Event>> liveData = new MutableLiveData<>();
         Query query = eventCollection.whereEqualTo("organizerId", organizerId);
+
+        return runQueryLiveData("getEventsOfOrganizerLiveData", query);
+    }
+
+    public LiveData<List<Event>> getEventsOfOrganizerLiveData(String organizerId, String facilityId) {
+        Query query = eventCollection
+                .whereEqualTo("organizerId", organizerId)
+                .whereEqualTo("facilityId", facilityId);
+
+        return runQueryLiveData("getEventsOfOrganizerLiveData", query);
+    }
+
+    public LiveData<List<Event>> getEventsOfFacilityLiveData(String facilityId) {
+        Query query = eventCollection.whereEqualTo("facilityId", facilityId);
+
+        return runQueryLiveData("getEventsOfFacilityLiveData", query);
+    }
+
+    private LiveData<List<Event>> runQueryLiveData(String methodName, Query query) {
+        MutableLiveData<List<Event>> liveData = new MutableLiveData<>();
 
         query.addSnapshotListener((querySnapshot, e) -> {
             if (e != null) {
-                Log.e(TAG, "getEventsOfOrganizerLiveData: listen failed", e);
+                Log.e(TAG, "runQueryLiveData: " + methodName + ": listen failed", e);
                 liveData.setValue(new ArrayList<>());
                 return;
             }
@@ -87,10 +110,10 @@ public class EventRepository {
                 for (int i = 0; i < events.size(); i++) {
                     events.get(i).setDocumentId(querySnapshot.getDocuments().get(i).getId());
                 }
-                Log.d(TAG, "getEventsOfOrganizerLiveData: success");
+                Log.d(TAG, "runQueryLiveData: " + methodName + ": success");
                 liveData.setValue(events);
             } else {
-                Log.d(TAG, "getEventsOfOrganizerLiveData: no events found");
+                Log.d(TAG, "runQueryLiveData: " + methodName + ": no documents found");
                 liveData.setValue(new ArrayList<>());
             }
         });
